Refactor UVEC package structure and update imports for consistency
