+++ UVEC/__init__.py
@@ -1,0 +1,3 @@
+from .__version__ import __version__, __title__
+
+
