+++ UVEC/__version__.py
