+++ UVEC/uvec_ten_dof_vehicle_2D/__init__.py
@@ -1,0 +1,3 @@
+from UVEC.copy_uvec import set_path_file
+
+UVEC_NAME = "uvec_ten_dof_vehicle_2D"
