+++ UVEC/uvec_ten_dof_vehicle_2D/newmark_solver.py
