+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = UVEC
-version = attr: uvec_ten_dof_vehicle_2D.__version__
-author = attr: uvec_ten_dof_vehicle_2D.__author__
+version = attr: UVEC.__version__
+author = attr: UVEC.__author__
 description = User defined VEhiCle models [UVEC]
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -12,8 +12,9 @@ classifiers =
 
 [options]
 packages =
-    uvec_ten_dof_vehicle_2D
-    uvec_two_dof_vehicle_2D
+    UVEC
+    UVEC.uvec_ten_dof_vehicle_2D
+    UVEC.uvec_two_dof_vehicle_2D
 include_package_data = True
 install_requires =
     numpy>=1.24
