+++ tests/test_moving_vehicle.py
@@ -1,10 +1,9 @@
 import json
 import matplotlib.pyplot as plt
 import numpy as np
-import numpy.typing as npt
 
-from uvec_ten_dof_vehicle_2D.uvec import uvec
-from uvec_ten_dof_vehicle_2D.newmark_solver import NewmarkExplicit, StaticSolver
+from UVEC.uvec_ten_dof_vehicle_2D.uvec import uvec
+from UVEC.uvec_ten_dof_vehicle_2D.newmark_solver import NewmarkExplicit, StaticSolver
 from tests.utils import UtilsFct
 from tests.analytical_solutions.moving_vehicle import TwoDofVehicle
