+++ tests/test_two_dof.py
@@ -1,8 +1,8 @@
 import json
 import numpy as np
 
-from uvec_two_dof_vehicle_2D.uvec import uvec
-from uvec_ten_dof_vehicle_2D.newmark_solver import NewmarkExplicit
+from UVEC.uvec_two_dof_vehicle_2D.uvec import uvec
+from UVEC.uvec_ten_dof_vehicle_2D.newmark_solver import NewmarkExplicit
 
 from tests.utils import UtilsFct
 from tests.analytical_solutions.moving_vehicle import TwoDofVehicle
