--- uvec_two_dof_vehicle_2D/__init__.py
@@ -1,1 +1,0 @@
-from .__version__ import __version__
