--- uvec_two_dof_vehicle_2D/__version__.py
@@ -1,3 +1,0 @@
-__title__ = "UVEC"
-__version__ = "1.0"
-__author__ = "STEM team"
